+++ src/api/generateFunctionSignatureGuard.ts
