+++ src/api/generateKeywordTypeGuardsForUnion.ts
@@ -10,26 +10,48 @@ import {
   isVoidKeyword,
 } from '../utils';
 
+/**
+ * Generates a type guard for the string keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateStringKeywordTypeGuard(kind: SyntaxKind): string {
   if (isStringKeyword(kind)) {
     return `(typeof value !== 'string')`;
   }
   return '';
 }
-
+/**
+ * Generates a type guard for the number keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateNumberKeywordTypeGuard(kind: SyntaxKind): string {
   if (isNumberKeyword(kind)) {
     return `(typeof value !== 'number')`;
   }
   return '';
 }
-
+/**
+ * Generates a type guard for the bigint keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateBigIntKeywordTypeGuard(kind: SyntaxKind): string {
   if (isBigIntKeyword(kind)) {
     return `(typeof value !== 'bigint')`;
   }
   return '';
 }
+/**
+ * Generates a type guard for the symbol keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 
 export function generateSymbolKeywordTypeGuard(kind: SyntaxKind): string {
   if (isSymbolKeyword(kind)) {
@@ -37,13 +59,21 @@ export function generateSymbolKeywordTypeGuard(kind: SyntaxKind): string {
   }
   return '';
 }
-
+/**
+ * Generates a type guard for the object keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateObjectKeywordTypeGuard(kind: SyntaxKind): string {
   if (isObjectKeyword(kind)) {
     return `(typeof value !== 'object' && value !== null)`;
   }
 }
-
+/**
+ * Generates a type guard for the 'any' keyword.
+ * NOTE: any keyword is not a type guard and typeguard should not be generated for it.
+ */
 export function generateAnyKeywordTypeGuard(kind: SyntaxKind) {
   //NOTE: any keyword is not a type guard and typeguard should not be generated for it
   return;
@@ -80,13 +110,23 @@ export function generateNeverKeywordTypeGuard(kind: SyntaxKind) {
   //NOTE: never keyword is not a type guard and typeguard should not be generated for it
   return;
 }
-
+/**
+ * Generates a type guard for the void keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateVoidKeywordTypeGuard(kind: SyntaxKind): string {
   if (isVoidKeyword(kind)) {
     return `(value !== undefined)`;
   }
 }
-
+/**
+ * Generates a type guard for the keyof keyword.
+ *
+ * @param {SyntaxKind} kind - The syntax kind to check.
+ * @returns {string} The generated type guard code.
+ */
 export function generateKeyofKeywordTypeGuard(kind: SyntaxKind): string {
   //TODO: fix this function to generate the correct type guard
   /*if(isKeyofKeyword(kind)) {
@@ -95,14 +135,24 @@ export function generateKeyofKeywordTypeGuard(kind: SyntaxKind): string {
     */
   return '';
 }
-
+/**
+ * Generates a type guard for the boolean keyword.
+ *
+ * @returns {string} The generated type guard code.
+ * @param type
+ */
 export function generateBooleanKeywordTypeGuard(type: SyntaxKind): string {
   if (isBooleanKeyword(type)) {
     return `(typeof value !== 'boolean')`;
   }
   return '';
 }
-
+/**
+ * Generates a type guard for the undefined keyword.
+ *
+ * @returns {string} The generated type guard code.
+ * @param type
+ */
 export function generateUndefinedKeywordTypeGuard(type: SyntaxKind): string {
   if (isUndefinedKeyword(type)) {
     return `(value !== undefined)`;
