+++ src/api/generateQualifiedNameTypeGuard.ts
