+++ src/api/index.ts
@@ -1,16 +1,22 @@
-export { generateInterfaceTypeGuard } from './generateInterfaceTypeGuard';
-export { generateIntersectionTypeGuard } from './generateIntersectionTypeGuard';
-export {
-  generateKeywordGuard,
-  generateKeywordGuardForType,
-} from './generateKeywordGuard';
-export { generatePropertyGuard } from './generatePropertyGuard';
-export { generateSingleInterfaceTypeGuard } from './generateSingleInterfaceTypeGuard';
-export { generateTypeReferenceGuard } from './generateTypeReferenceGuard';
-export { generateUnionTypeGuard } from './generateUnionTypeGuard';
-export {
-  generateEnumTypeGuards,
-  handleEnumIntersection,
-  generateEnumTypeGuard,
-} from './generateEnumTypeGuard';
-export { handleHeritageClauses } from './handleHeritageClauses';
+export * from './generateInterfaceTypeGuard';
+export * from './generateEnumTypeGuard';
+export * from './generateFunctionSignatureGuard';
+export * from './generateInterfaceTypeGuard';
+export * from './generateIntersectionTypeGuard';
+export * from './generateIntersectionTypeGuardForType';
+export * from './generateKeywordGuard';
+export * from './generateKeywordTypeGuardsForUnion';
+export * from './generateLiteralTypeTypeGuard';
+export * from './generateOptionalPropertyTypeGuard';
+export * from './generatePropertyGuard';
+export * from './generateQualifiedNameTypeGuard';
+export * from './generateReadonlyPropertyTypeGuard';
+export * from './generateSingleInterfaceTypeGuard';
+export * from './generateTypeLiteralTypeGuard';
+export * from './generateTypeReferenceGuard';
+export * from './generateUnionTypeGuard';
+export * from './handleHeritageClauses';
+
+export * from './generateUnionTypeGuardForIntersection';
+
+export * from './generateTypeTypeGuard';
