--- src/generator/generateInterfaceTypeGuard.ts
@@ -1,1 +1,0 @@
-// Generate type guards for nested interfaces/types
